Fix for back camera showing briefly when the type property is set to 'front'.
The first time an app is run that uses this component with the type set to 'front', the back camera shows briefly before the component obeys the type property and switches to the front camera. The presetCamera should be set to the proper camera type set in the props when the component is created.